Add copies of current raw HTML from site
This adds a folder that contains all the raw HTML/etc. found in the
"Fullwidth Code" block of every existing HNN website page, except for a
few. The only exceptions are

- event-feedback.html
- pre-event-survey.html
- workshop-registration.html

all of which simply embed a Google Form and do not need to have their
content recreated in the Redesign, and

- news-and-events.html

which is unique in that it is the only page that appears to have been
assembled using actual Wordpress itself, instead of raw HTML/etc.