Add MANUAL run of opt nbs, deleting some output
NOTE: For this commit, I did the following:

1. Local, external runs of the optimize notebooks (i.e. not using
execute-notebooks).
2. For each notebook, deleted the text output of ONLY the "main" cell
that does the optimization, which consists of almost a megabyte (!) of
text.
3. For the JSON file corresponding to each notebook, manually changed
the `full_executed` flag to true.
4. Added the new `output_*` folders to git.
5. Then did the usual website build make recipe, which updated the rest
of the (new) output into the JSON and HTML files.

This process only needs to be done when the Optimize notebooks need to
be updated, which should be rarely.